requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
asttokens==2.0.5
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
backcall==0.2.0
certifi==2022.6.15
charset-normalizer==2.0.12
colorama==0.4.5
cycler==0.11.0
debugpy==1.6.0
decorator==5.1.1
entrypoints==0.4
executing==0.8.3
filelock==3.7.1
fonttools==4.33.3
frozenlist==1.3.0
huggingface-hub==0.8.1
idna==3.3
ipykernel==6.15.0
ipython==8.4.0
jedi==0.18.1
jupyter-client==7.3.4
jupyter-core==4.10.0
kiwisolver==1.4.3
matplotlib==3.5.2
matplotlib-inline==0.1.3
multidict==6.0.2
nest-asyncio==1.5.5
numpy==1.22.4
packaging==21.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.1.1
prompt-toolkit==3.0.29
psutil==5.9.1
pure-eval==0.2.2
Pygments==2.12.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
regex==2022.6.2
requests==2.28.0
six==1.16.0
stack-data==0.3.0
timm==0.5.4
tokenizers==0.12.1
torch==1.11.0
torchaudio==0.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.20.0
typing_extensions==4.2.0
urllib3==1.26.9
wcwidth==0.2.5
yarl==1.7.2
